+++ frontend/src/components/ReviewModal/index.tsx
@@ -1,0 +1,3 @@
+import ReviewModal from '~/components/ReviewModal/ReviewModal';
+
+export default ReviewModal;
